Fix material highlighting not resetting
